lab4_multiagent_mcp/pyproject.toml
[project]
name = "lab4-multiagent-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
]
